UI of second page

+++ data/lib/fitness/entities/api/goal_api_dto.dart
@@ -9,7 +9,7 @@ class GoalApiDto {
   final String title;
   final String description;
   @JsonKey(name: 'duration')
-  final double durationSeconds;
+  final int durationSeconds;
   @JsonKey(name: 'burn_calories')
   final int burnCalories;
